Multilanguage Like
I've tested with multiple languages, resolutions, users and locations
and seems that the coreSpriteLikeHeart class never changes.
When the image is not liked yet, the class is coreSpriteLikeHeartOpen,
when the image is already liked the class changes to
coreSpriteLikeHeartFull.
This should fix the language/locale issue with the like function, but
the comment function must be reviewed.