Make like_util work independent of 'use_new_styles'
With the #33 I realized that there’s a JS variable that contains all
data that we need and that the UI of Instagram Web differs from users
(use_new_styles flag).

Changed the XPATH/DOM references for JS functions to retrieve the data,
independent of ‘use_new_styles’.

Also made a change on InstaPy.py that forces the Chrome to run with the
“en-US” location to avoid the Locale/Language problem with the like
button and the comment placeholder.