comment_util changes with 'textarea' support
This would explain why some people were having problems with the bot skipping commenting and why some were having it crash. This should now check for text area, if it doesn't exist it will check for the input box instead. If it cannot find either the bot will leave a warning message the commenting might not have worked.