Fixed corrupted cookie
If the cookie associated with a login becomes "bad" for any reason, InstaPy previously did not have the ability to catch this. The most common reason for a cookie failing to work, is the user changing their password.

login_util.py will check if a cookie is present, and if so, load it. With these changes, InstaPy will ensure the user is successfully logged in by checking if when loading www.instagram.com the user still sees a "Log in" button. If so, the cookie is assumed to be "bad" and re-created.