Fix an issue that cause duplicated console info
Basically similar issue as described from PR #1255 but the fix he put in there wasn't real fix. the issue was the console_handler get duplicated. There is no need check if a logger already existed or not, consider the logger is singleton and always be the same object regardless what. reference from the official logging module page `Loggers have the following attributes and methods. Note that Loggers are never instantiated directly, but always through the module-level function logging.getLogger(name). Multiple calls to getLogger() with the same name will always return a reference to the same Logger object.`