Revised `like_by_users` feature to speed up the process
Now when a user's relationship bounds is matched for once, it will not be checked in the next posts of that same user (by navigating to the profile page in each post..) again.
Similarly, if a user's relationship bounds is not matched with the range desired, then it will simnply skip that user without seeing further posts.
Also, I have moved the process of getting follower and following counts into a new function for ease of maintenance and to make usable in anwhere else, too, e.g., right from quickstart scripts...