Sync to latest version on github.
session.set_skip_users(skip_private=True, skip_no_profile_pic=True, min_media=25)
- Skip private profile
- Skip users with less posts than a specific number
- Skip user without a profile pic. (not fully tested yet)

Fix problem in clarifi_util.py, changed the part of the code to retrieve the image url, using graphql.shortcode_media.display_url.

Work in progress

Work in progress

updated README.md

changed link with format()

added max_media parameter

Not tested yet

Minor changes
changed fault value of media_c to None, check if is not None, add in log media_c

Possible bug fix, changed url = None to url = "", this may prevent crash, if page doesn't exist (maybe user non existing anymore)

Add the possibility to set no limit on the max_media.

Updated README.md

Little update to README.md

Work in progress, not tested yet.

Possible bug fix in get_relationship_counts but not sure. (happens when the page opened is of a not existing user)

instapy.py:

change in follow_likers:
validate first users, add on a list a given ammout of valid users (or how may valid users there is, if is less than the given ammount)
then follow.
I'm not sure, if the count of followed is right, to take a break after followed is more than relax_point.
Added some debug info in the logger, just temporary.

util.py:

not sure if is right. My attempt is to detect as invalid user if the user is not existing anymore (or just a wrong page)

Skip also users already folled more than n times

Work in progress,
	for now I’m working only on follow_likers.

moved min_media and max_media in set_relationship_bounds

add parameter max_relationship_ratio in set_relationship_bounds, to delimit also with a maximum ratio.

For now, this is a work in progress, so the new behavior is the default, but I’ll add a parameter to use the previous version.

Basically, first I validate the users, when i get the likers, from the likes grabbed, I add this people, in a list.
If the grabbed people is less than the desired number, pick again the likes, ad add the new ones grabbed to the list.
This process cannot be done forever, so i decided to add maximum attempts, I chose to use 3 as maximum.

Another thing, I added, a very little delay, between user validation. Now, for testing purposes is set to a very low number.
The idea, is simulate, the user checking the profile. So basically, stay a little bit on this profile, even if is not valid, and then move on to the next one

Because, we collect a large number of people to follow, I added also a delay between the follows.

For now is all, folks :)

Work in progress,
for now I’m working only on follow_likers.

moved min_media and max_media in set_relationship_bounds

add parameter max_relationship_ratio in set_relationship_bounds, to delimit also with a maximum ratio.

For now, this is a work in progress, so the new behavior is the default, but I’ll add a parameter to use the previous version.

Basically, first I validate the users, when i get the likers, from the likes grabbed, I add this people, in a list.
If the grabbed people is less than the desired number, pick again the likes, ad add the new ones grabbed to the list.
This process cannot be done forever, so i decided to add maximum attempts, I chose to use 3 as maximum.

Another thing, I added, a very little delay, between user validation. Now, for testing purposes is set to a very low number.
The idea, is simulate, the user checking the profile. So basically, stay a little bit on this profile, even if is not valid, and then move on to the next one

Because, we collect a large number of people to follow, I added also a delay between the follows.

For now is all, folks :)

Solved conflicts thanks to @uluQulu

Updated to latest code base

Restored my changes

Work in progress, changed parameters name in private_users and no_profile_pic, moved min_media and max_media in set_skip_users, added some minor change in follow_likers

update clarifai_util as master

Merged two functions to get if is a private profile, in one. Not sure if is all ok, still a work in progress and need a test.