Turn off plyer toast notifications for python 2 in Windows out of a bug with 'ticker'
That bug prints an annoying inner exception message which isn't thrown to be catched, it just prints it all the time.
That issue is fixed in the new beta version of plyer but it is not pushed to its stable version at PyPI yet.
So disabling it for now is better.