Increase code strenght at checking notification offer popup block
`Turn On` is a button and it can be anywhere on a popup not-loaded page.
`Turn on Notifications` is a dialog name.
By code strenght I mean locating the right element we want to refer to and not any other one.
It is the same way I have offered in my initial comment.