Major upgrades and refactor in get_following_status function.
Mainly, it has a clearer and wiser design.
E.g. it now returns the exact following state rather than True, False, None - this change was due to provide flexibilty and clear code.
And it now checks if the page is available before trying to find the following status.
Also at that stage, finding if the user has changed username is being run accordingly in this function.
The difference is lots of precious time from the extra queries happened.