Added location to image_text
Added the location (if any) of an image to the image_text so the mandatory words check also checks the location. 

This is useful when targeting images with a certain hashtag that where taking in a certain location. Sometimes the user tags the location, sometimes the location is in the description. Both should be included.