Fix Follow-Likers feature which couldn't fetch likers properly
Issue raised by @opexa at #3573.
The problem occurs when expected text has a comma in it.
E.g. "100 others" works but "3,872 others" will fail.
It's due to the usage of re module in there with its \d+ key it matches decimals and does not expect a comma in decimals.

re is an overuse in there, I used plain string matching instead.