Move icons folder (with its content) to instapy package [folder]
When instapy package is installed through pip then icons folder will reside inside the instapy folder.
Without this change, it would reside in a separate folder at the same level with instapy package which is not good.

With setuptools it is impossible to make it work otherwise (it was possible with old distutils).