Update unfollow()
Managing InstapyFollowed and nonFollowers parameters = True in unfollow()

When passing both parameters InstapyFollowed and nonFollowers as True to unfollow(), the second is ignored as the condition InstapyFollowed comes first. The suggestion below would ignore InstapyFollowed and base the following actions on nonFollowers (which is itself based on the InstapyFollowed list minus users that followed back at any time).